+++ api/src/main/java/io/github/sds100/keymapper/api/PauseMappingsBroadcastReceiver.kt
@@ -8,7 +8,7 @@ import io.github.sds100.keymapper.base.keymaps.PauseKeyMapsUseCase
 import io.github.sds100.keymapper.common.utils.firstBlocking
 import javax.inject.Inject
 
-// DON'T MOVE THIS CLASS TO A DIFFERENT PACKAGE BECAUSE IT BREAKS THE API
+// DON'T MOVE THIS CLASS TO A DIFFERENT PACKAGE OR RENAME BECAUSE IT BREAKS THE API
 @AndroidEntryPoint
 class PauseMappingsBroadcastReceiver : BroadcastReceiver() {