+++ data/src/main/java/io/github/sds100/keymapper/data/repositories/KeyMapRepository.kt
@@ -21,5 +21,6 @@ interface KeyMapRepository {
     fun disableById(vararg uid: String)
     fun enableByGroup(groupUid: String?)
     fun disableByGroup(groupUid: String?)
+    fun toggleById(vararg uid: String)
     fun moveToGroup(groupUid: String?, vararg uid: String)
 }